# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FORTRESS AI Cloud Defender & Security Scanner API",
    "version": "3.5.0",
    "description": "Always-Active Military-Grade API Security Wall, AI Threat Scanner, and Autonomous Virtual Patching Engine powered by Google Gemini.\n\n### Authentication\nMost endpoints require an authorized Vault Key or Master Passphrase:\n- Header: `x-vault-key: <your_key>`\n- Header: `Authorization: Bearer <your_key>`\n- Header: `x-vault-pass: <master_passphrase>`\n\nPublic endpoints: `/health`, `/docs`, `/openapi.json`.",
    "contact": {
      "name": "FORTRESS Security Engineering",
      "url": "https://github.com/alexhack235-code/AI-Security-Scanner-API"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://ai-security-scanner-api.vercel.app",
      "description": "Cloud Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "VaultKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-vault-key",
        "description": "Client Vault Key issued by Keymaster"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "VaultKey",
        "description": "Client Vault Key provided as Bearer token"
      },
      "MasterPassAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-vault-pass",
        "description": "Master Vault Administrative Passphrase"
      }
    },
    "schemas": {
      "DefendRequest": {
        "type": "object",
        "required": ["method", "path"],
        "properties": {
          "method": {
            "type": "string",
            "example": "POST",
            "description": "HTTP Method of incoming request"
          },
          "path": {
            "type": "string",
            "example": "/api/users/update",
            "description": "Route path of incoming request"
          },
          "headers": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "example": { "content-type": "application/json", "user-agent": "Mozilla/5.0" }
          },
          "query": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "example": { "id": "105" }
          },
          "body": {
            "type": "object",
            "description": "Request body payload to inspect and sanitize",
            "example": { "bio": "Security researcher <script>alert(1)</script>" }
          },
          "ip": {
            "type": "string",
            "example": "198.51.100.42",
            "description": "Client IP address for telemetry and auto-jailing"
          },
          "userAgent": {
            "type": "string",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)"
          }
        }
      },
      "DefendResponse": {
        "type": "object",
        "properties": {
          "verdict": { "type": "string", "example": "ALLOWED" },
          "action": { "type": "string", "example": "FORWARD_REQUEST" },
          "threatLevel": { "type": "string", "example": "LOW" },
          "anomalyScore": { "type": "number", "example": 0.05 },
          "risk": { "type": "string", "example": "LOW" },
          "zero_vulnerability_guarantee": { "type": "boolean", "example": true },
          "clean_data": { "type": "object" },
          "applied_hotpatches": { "type": "array", "items": { "type": "string" } },
          "details": { "type": "string", "example": "Traffic verified clean by FORTRESS AI Shield." }
        }
      },
...
      "CodeScanRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Source code text to audit with Gemini Neural Engine",
            "example": "app.get('/user', (req, res) => { db.query('SELECT * FROM users WHERE id = ' + req.query.id); });"
          },
          "filename": {
            "type": "string",
            "example": "userService.js"
          }
        }
      },
...
      "StandardError": {
        "type": "object",
        "properties": {
          "fortress_status": { "type": "string", "example": "REJECTED" },
          "threat_level": { "type": "string", "example": "HIGH" },
          "reason": { "type": "string", "example": "Malicious payload detected." },
          "action": { "type": "string", "example": "DROP_CONNECTION" }
        }
      }
    }
  },
  "security": [
    { "VaultKeyAuth": [] },
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/defend": {
      "post": {
        "tags": ["Cloud Defender & WAF"],
        "summary": "Autonomous Multi-Layer Defense & AI Cognitive Sanitizer",
        "description": "The primary defense endpoint. Inspects HTTP requests against 12 Fast-Kill WAF rules, applies active virtual hotpatches, detects anomalies, and sanitizes payloads via Google Gemini.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DefendRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request evaluated clean or sanitized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DefendResponse" }
              }
            }
          },
          "403": {
            "description": "Request blocked by Fast-Kill WAF or Virtual Patch",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardError" }
              }
            }
          }
        }
      }
    },
...
    "/api/scan": {
      "post": {
        "tags": ["Vulnerability Scanners"],
        "summary": "Gemini Deep Neural SAST Code Auditor",
        "description": "Submits source code snippets or files for automated Static Application Security Testing (SAST) powered by Google Gemini AI. Identifies OWASP Top 10 flaws, CWE mappings, and remediation code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CodeScanRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SAST code audit complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vulnerabilities_found": { "type": "integer", "example": 1 },
                    "risk_level": { "type": "string", "example": "HIGH" },
                    "findings": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vault/verify": {
      "post": {
        "tags": ["Vault Keymaster & Auth"],
        "summary": "Verify Client Vault Key or Master Passphrase",
        "description": "Tests if a provided Vault Key or Master Passphrase is valid and returns remaining quota and role information.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": { "type": "string", "example": "fortress_live_sec_..." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Key verified authentic" },
          "401": { "description": "Invalid or revoked key" }
        }
      }
    },
    "/api/vault/keys": {
      "get": {
        "tags": ["Vault Keymaster & Auth"],
        "summary": "List Active Client Vault Keys",
        "description": "Returns all issued API keys, usage statistics, and remaining quotas. Requires Master Admin role.",
        "security": [{ "MasterPassAuth": [] }],
        "responses": {
          "200": { "description": "List of active client keys" },
          "403": { "description": "Requires Master Vault Passphrase" }
        }
      },
      "post": {
        "tags": ["Vault Keymaster & Auth"],
        "summary": "Issue New Client Vault Key",
        "description": "Generates a cryptographically secure Vault Key for a client or organization. Requires Master Admin role.",
        "security": [{ "MasterPassAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Acme Corp Production" },
                  "quota": { "type": "integer", "example": 5000 },
                  "role": { "type": "string", "enum": ["CLIENT", "DEVELOPER", "ADMIN"], "example": "CLIENT" },
                  "notes": { "type": "string", "example": "Production checkout API" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Key created successfully" },
          "403": { "description": "Requires Master Vault Passphrase" }
        }
      }
    },
...
    "/api/admin/handshake/issue": {
      "post": {
        "tags": ["Ephemeral Admin Handshake"],
        "summary": "Issue Time-Bombed One-Way Admin Ticket",
        "description": "Generates a single-use ephemeral ticket that self-destructs after 20 to 60 seconds. Prevents session hijacking and persistent backdoors.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ttl_seconds": { "type": "integer", "default": 60, "example": 30 },
                  "scope": { "type": "string", "default": "ADMIN_TELEMETRY" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Ephemeral ticket issued" }
        }
      }
    },
    "/api/admin/handshake/claim": {
      "post": {
        "tags": ["Ephemeral Admin Handshake"],
        "summary": "Claim Ephemeral Admin Ticket",
        "description": "Consumes the one-way ticket to execute an authorized administrative telemetry fetch. The ticket is immediately invalidated upon first use.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": { "type": "string", "example": "eph_9f81a7b..." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Admin telemetry returned and ticket incinerated" },
          "403": { "description": "Ticket expired or already claimed" }
        }
      }
    },
...
  }
}